# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateUserDTO:
    properties:
      cartID:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  dto.UpdateUserDTO:
    properties:
      cartID:
        type: string
      name:
        type: string
    type: object
  v1.failure:
    properties:
      error:
        $ref: '#/definitions/v1.failureInfo'
    type: object
  v1.failureInfo:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request body
        type: string
    type: object
  v1.success:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: paw1a@yandex.ru
    name: API Support
    url: https://t.me/paw1a
  description: This is simple api of e-commerce shop
  title: E-commerce API
  version: "1.0"
paths:
...
  /admins/users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get all users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Create user
      tags:
      - admin-users
  /admins/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Delete user
      tags:
      - admin-users
    get:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get user by id
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: user update fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Update user
      tags:
      - admin-users
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"